Asset, Transaction value 자료형 Long으로 수정. close#128, close#130

+++ backend/src/main/java/com/back/domain/asset/entity/Asset.java
@@ -23,7 +23,7 @@ public class Asset extends BaseEntity {
     @Enumerated(EnumType.STRING)
     AssetType assetType; // 자산 유형
 
-    int assetValue; // 자산 가치
+    Long assetValue; // 자산 가치
 
     // int id(PK) -> BaseEntity
     // LocalDateTime created_at -> BaseEntity
