+++ backend/src/main/java/com/back/domain/transactions/dto/CreateTransactionRequestDto.java
@@ -3,11 +3,11 @@ package com.back.domain.transactions.dto;
 public record CreateTransactionRequestDto(
         int assetId,
         String type,
-        int amount,
+        Long amount,
         String content,
         String date
 ) {
-    public CreateTransactionRequestDto(int assetId, String type, int amount, String content, String date) {
+    public CreateTransactionRequestDto(int assetId, String type, Long amount, String content, String date) {
         this.assetId = assetId;
         this.type = type;
         this.amount = amount;
